Delegate graceful shutdown to a caller-provided quit signal
This commit preserves the existing `serve_fcgid` function, but it's no longer
meant as the main entry point; the new `serve_fcgid_with_graceful_shutdown`
function takes that place.

This comes out of getting a bit more experience with assembling a serious axum
app, which actually has a variety of shutdown concerns it wants to manage.