Example app: take control of graceful shutdown handling
This bears a closer resemblance to what I ended up doing in my serious app,
where the app code owns the shutdown signal and passes a resulting signal to the
serve futures.

(Actually in the real app we use a `tokio_util::sync::CancellationToken`, pass
its `.cancelled_owned()` to the serve future, and spawn a task to listen for
kill signals and call `.cancel()` on the token. More indirect, but it let me
broadcast the shutdown to some other non-serving tasks.)